backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Fast API Backend for Python Learning Club website"
authors = ["steven <msycomtw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.3"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
sqlmodel = "^0.0.18"
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
